requirements.txt
flask
flask_sqlalchemy
flask_wtf
flask-login
flask_bcrypt
wtforms
wtforms[email]
Pillow
